refactor : 에러코드 디렉토리 생성
